refactor: no cheering team 객체 상수로 분리
