+++ src/components/common/account/favorite-team-item.tsx
@@ -5,7 +5,7 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
-import { NO_CHEERING_TEAM_PK } from '@/lib/constants';
+import { NO_CHEERING_TEAM_PK } from '@/lib/constants/noCheeringTeam';
 
 export default function FavoriteTeamItem({
 	team,
