+++ src/components/common/account/favorite-team-section.tsx
@@ -5,7 +5,7 @@ import { TeamDto } from '@/services/apis/team/dto';
 import dynamic from 'next/dynamic';
 import FavoriteTeamItem from './favorite-team-item';
 import SelectSection from './select-section';
-import { NO_CHEERING_TEAM_PK } from '@/lib/constants';
+import { NO_CHEERING_TEAM, NO_CHEERING_TEAM_PK } from '@/lib/constants/noCheeringTeam';
 import FavoriteTeamHeader from './favorite-team-header';
 
 // dnd-kit 컴포넌트 hydration mismatch 가능성 존재 -> ssr 비활성화
@@ -44,21 +44,7 @@ export default function FavoriteTeamSection({
 
 	useEffect(() => {
 		if (initialTeams) {
-			const teams =
-				initialTeams.length === 0
-					? [
-							{
-								pk: -1,
-								nameKr: '응원팀이 없어요',
-								nameEn: 'no cheering team',
-								logoUrl: '/ban.svg',
-								leaguePk: -1,
-								leagueNameKr: '응원팀이 없어요',
-								leagueNameEn: 'no cheering team',
-								leagueLogoUrl: '/ban.svg',
-							},
-						]
-					: initialTeams;
+			const teams = initialTeams.length === 0 ? [NO_CHEERING_TEAM] : initialTeams;
 			setFavoriteTeams(teams);
 		}
 	}, [initialTeams]);
