--- src/lib/constants/index.ts
@@ -1,1 +1,0 @@
-export const NO_CHEERING_TEAM_PK = -1;
