+++ src/lib/constants/noCheeringTeam.ts
@@ -1,0 +1,12 @@
+export const NO_CHEERING_TEAM_PK = -1;
+
+export const NO_CHEERING_TEAM = {
+	pk: NO_CHEERING_TEAM_PK,
+	nameKr: '응원팀이 없어요',
+	nameEn: 'no cheering team',
+	logoUrl: '/ban.svg',
+	leaguePk: -1,
+	leagueNameKr: '응원팀이 없어요',
+	leagueNameEn: 'no cheering team',
+	leagueLogoUrl: '/ban.svg',
+};
